feat: Optimized the sorting function of the librarybar
