+++ src/renderer/src/components/Librarybar/main.tsx
@@ -1,6 +1,7 @@
 import { Button } from '@ui/button'
 import { Input } from '@ui/input'
 import { Nav } from '@ui/nav'
+import { Popover, PopoverContent, PopoverTrigger } from '@ui/popover'
 import {
   Select,
   SelectContent,
@@ -12,6 +13,7 @@ import {
 } from '@ui/select'
 import { Tooltip, TooltipContent, TooltipTrigger } from '@ui/tooltip'
 import { isEqual } from 'lodash'
+import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useConfigState } from '~/hooks'
 import { cn } from '~/utils'
@@ -29,6 +31,16 @@ export function Librarybar(): JSX.Element {
   const filter = useFilterStore((state) => state.filter)
   const toggleFilterMenu = useFilterStore((state) => state.toggleFilterMenu)
   const clearFilter = useFilterStore((state) => state.clearFilter)
+
+  const [isSortMenuOpen, setIsSortMenuOpen] = useState(false)
+  const [by, setBy] = useConfigState('game.gameList.sort.by')
+  const [order, setOrder] = useConfigState('game.gameList.sort.order')
+  const toggleSortMenu = (): void => {
+    setIsSortMenuOpen(!isSortMenuOpen)
+  }
+  const toggleOrder = (): void => {
+    setOrder(order === 'asc' ? 'desc' : 'asc')
+  }
 
   console.warn(`[DEBUG] Librarybar`)
 
@@ -103,26 +115,86 @@ export function Librarybar(): JSX.Element {
             )}
           </div>
         </div>
-        <div className={cn('pr-3')}>
-          <Select value={selectedGroup} onValueChange={setSelectedGroup}>
-            <SelectTrigger className="pr-3">
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectGroup>
-                <SelectLabel>{t('librarybar.groupBy')}</SelectLabel>
-                <SelectItem value="none">{t('librarybar.groups.none')}</SelectItem>
-                <SelectItem value="collection">{t('librarybar.groups.collection')}</SelectItem>
-                <SelectItem value="metadata.developers">
-                  {t('librarybar.groups.developers')}
-                </SelectItem>
-                <SelectItem value="metadata.genres">{t('librarybar.groups.genres')}</SelectItem>
-                <SelectItem value="record.playStatus">
-                  {t('librarybar.groups.playStatus')}
-                </SelectItem>
-              </SelectGroup>
-            </SelectContent>
-          </Select>
+        <div className={cn('flex flex-row gap-2 pr-3 items-center')}>
+          <div className={cn('flex-1 min-w-0')}>
+            <Select value={selectedGroup} onValueChange={setSelectedGroup}>
+              <SelectTrigger className="pr-3">
+                <SelectValue />
+              </SelectTrigger>
+              <SelectContent>
+                <SelectGroup>
+                  <SelectLabel>{t('librarybar.groupBy')}</SelectLabel>
+                  <SelectItem value="none">{t('librarybar.groups.none')}</SelectItem>
+                  <SelectItem value="collection">{t('librarybar.groups.collection')}</SelectItem>
+                  <SelectItem value="metadata.developers">
+                    {t('librarybar.groups.developers')}
+                  </SelectItem>
+                  <SelectItem value="metadata.genres">{t('librarybar.groups.genres')}</SelectItem>
+                  <SelectItem value="record.playStatus">
+                    {t('librarybar.groups.playStatus')}
+                  </SelectItem>
+                </SelectGroup>
+              </SelectContent>
+            </Select>
+          </div>
+          <div>
+            <Popover open={isSortMenuOpen}>
+              <Tooltip>
+                <PopoverTrigger>
+                  <TooltipTrigger asChild>
+                    <Button onClick={toggleSortMenu} variant="outline" size={'icon'}>
+                      <span className={cn('icon-[mdi--sort] w-5 h-5')}></span>
+                    </Button>
+                  </TooltipTrigger>
+                </PopoverTrigger>
+                <TooltipContent side="right">{t('list.all.sortBy')}</TooltipContent>
+              </Tooltip>
+              <PopoverContent side="right">
+                <div className={cn('flex flex-row gap-5 items-center justify-center')}>
+                  <div className={cn('flex flex-row gap-1 items-center justify-center')}>
+                    <div className={cn('text-sm whitespace-nowrap')}>{t('list.all.sortBy')}：</div>
+                    <Select value={by} onValueChange={setBy} defaultValue="name">
+                      <SelectTrigger className={cn('w-[120px] h-[26px] text-xs')}>
+                        <SelectValue placeholder="Select a fruit" className={cn('text-xs')} />
+                      </SelectTrigger>
+                      <SelectContent>
+                        <SelectGroup>
+                          <SelectLabel>{t('list.all.sortBy')}：</SelectLabel>
+                          <SelectItem value="metadata.name">
+                            {t('list.all.sortOptions.name')}
+                          </SelectItem>
+                          <SelectItem value="metadata.releaseDate">
+                            {t('list.all.sortOptions.releaseDate')}
+                          </SelectItem>
+                          <SelectItem value="record.lastRunDate">
+                            {t('list.all.sortOptions.lastRunDate')}
+                          </SelectItem>
+                          <SelectItem value="record.addDate">
+                            {t('list.all.sortOptions.addDate')}
+                          </SelectItem>
+                          <SelectItem value="record.playTime">
+                            {t('list.all.sortOptions.playTime')}
+                          </SelectItem>
+                        </SelectGroup>
+                      </SelectContent>
+                    </Select>
+                  </div>
+                  <Button
+                    variant={'outline'}
+                    size={'icon'}
+                    className={cn('h-[26px] w-[26px] -ml-3')}
+                    onClick={toggleOrder}
+                  >
+                    {order === 'asc' ? (
+                      <span className={cn('icon-[mdi--arrow-up] w-4 h-4')}></span>
+                    ) : (
+                      <span className={cn('icon-[mdi--arrow-down] w-4 h-4')}></span>
+                    )}
+                  </Button>
+                </div>
+              </PopoverContent>
+            </Popover>
+          </div>
         </div>
         <GameList query={query} selectedGroup={selectedGroup} />
       </div>
